backend/requirements.txt
# Core Framework
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Security & Authentication
bcrypt==4.1.1
PyJWT==2.8.0
cryptography==41.0.7

# Validation & Serialization
marshmallow==3.20.1
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0

# Environment & Configuration
python-dotenv==1.0.0

# Date/Time handling
python-dateutil==2.8.2

# API Documentation
flask-swagger-ui==4.11.1

# Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0

# Code Quality
black==23.12.0
flake8==6.1.0
pylint==3.0.3

# Utilities
click==8.1.7
Werkzeug==3.0.1
